Updated the 2nd demo. Now it's more colorfull, with icons and stuff...
I also updated the help displayed on that demo, and added an help browser,
to demostrate how to use 2 different mdi clients (with a non mdi client displayed 
when the application starts). 

This demo is almost a very functional application. I will try to add open/save/search
support, and it's a fine looking editor :)

The icons are from Qt 4.1, besides fileclose.png which is part of KDE's Cristal icon set. 
All free.