= core library =  * added a changelog to the doxygen documentation  * qmdihost::updateGUI can be called without any parameter to update itself
= plugin demo =
 * configuration dialog dialog is created on demand (on the first call to the "config" action
 * model has been udpated, now it really works, and displays the list of plugins available
 * config dialog is not functional, but it displays something

== todo ==
The configuration dialog will be splited into 2 different dialogs: one to enable/disable and query the list of
plugins, the second will display the configuration dialog of each plugin.

The code in the plugin demo is a mess, and needs to be reviewed and refactored once the basic idea is out and running.