iPlugin: function to get the project manager
There are some occasions in which from the plugin  you need to interact
with the manager itself. You can typast your way up - of use this
function.

Right now - it just casts up. However - I might add a link to the
manager indede the plugin, so this function will be stable api.