New sub system: config
This new API lets you define a schema for configuration (an itemm call
"foo" can contain a number, and default is 42, "bar" is a string...
etc).

The values can be saved inot disk and later on restored.

GUI for modifying the configuration is automatically generated.

Code contains unit tests for the new API (GUI is not tested).