qmdiConfigItem/IPlugin: read the config from QSettings
IPlugin now reads the config items from QSettings. We also gained the
ability to have configs which are not user editable, meaning - no UI
will be generated for them. Instead - in my examples, they are deducted
from UI state.

This fixes the issue where we don't know how to save the config file.
Instead we ue the standard QSettings system - which is set by the
application.