demo2: fix and modernize
For some reason, the code used QTabWidget instead qmdiTabWidget, which
messed up all the features.

Also - on my setup (I use the online installer) the help was not
loading. As a hack, to show the library, I am at least trying to load
the QtCreator help. At least something works.