qmdiClient: support for context menus
When you right click a qmdiServer, a menu (with close, close all and
close all but...) is opened. From now on, you can define in the clients
which actions will be merged at the end of this menu.

Added this to the qmdiEditor in the second example and PluginManager
example.