PluginManager: single new document command
In a previous design, I intended each plugin to have a set of "new"
items, then from the first icon on the tab you would choose which file
to create.. this is not something people expect today.

I changed this to a "new file command" and let the application handle
this.

Found the bug which made we always had an empty tab. It was a non empty
QTabWidget in the designer.