PluginManager: sync settings after load
In case any pluing modified the config - lets save this. This occurs in
QtEdit4 when the "last update check" is updated via a plugin. If the
plugin saves during this stage, no files have been loaded yet, which
results in the loaded file list being cleared upon the next application
launch.