PluginManager: work in hidden gui mode
Up until now, if you hide the GUI, the actions you added to the menu
were not available. However, you could still acess them via the command
pallete (in qtedit4) - if that would work.

Problem is that the actions are only in the menus/toolbars. And since
those are hidden - inaccesible. The solution, is to add the actions to
the main window. Now shortcuts work also when the menu is hidden.

I have some actions duplicated (I can see it in "save").