qmdiServer: onMdiSelected callback
Some applications need to be motified when a new mdi client is selected
(for example to update the window title, or update status of some
commands).

The way this is exposed - it is via a function on the mdi server.

Before something like this was done:

```
connect(tabWidget, &QTabWidget::currentChanged, this,
&PluginManager::updateActionsStatus);
```

Now this code should be used:
```
mdiServer->setOnMdiSelected([this](qmdiClient *, int) {
updateActionsStatus(); });
```

Added this implementation on qmdiTabWidget.