qmdiConfigDialog: add support for buttons
You can create buttons as config. They save nothing (yet, there is a bug
and data is saved in the config file).

When the user clicks that button, the click is sent to the global dialog
events - and your application can intercept it, and do work uppon it.

This kinda breaks the way this class works, as it deals with UI and not
pure config.