PluginManager: fix closing of docks
Previously, when closing a dock in minimized mode, the UI was not
properly saved/restored.

The problem was that in minimized mode, I locked the docks from being
closable. So, the actions were disabled. Which is logical.

BUT, when saving, there was an issue internally in Qt : a non closable
dock was closed - all configuration must be broken. Do not save/restore.

New problem:
I need to merge toolbars even in minimized mode - as the use might re-
enable it from the toolbars menu, even in minimized mode.