qmdiHost: remove locking of menus/toolbars
The assumption while writing this code "back then" was that updating the
UI is too  consumimg. I think this is wrong.

Removing this merges menus/toolbars also in minimized mode. Which is
good, as the globals actions may be avialable by the command palette in
qtedit4, or the application menu soon to come.