qmdiClient, PluginManager: support for random state saving
In preparation for https://github.com/diegoiast/qtedit4/issues/19, this
modification makes possible for each qmdiClient save a different state
to the config.

This enables the editor to save row/col/zoom, and the image viewer
scroll+zoom. In the future, the text editor will be able to store also
breakpoint info or per editor configuration (theme, highlighting,
indenter whatever).