qmdiClient: when closing, delete client (#30)
The real problem I found was the `o->deleteLater()`. I changed this to
`delete o`, which is equivalent to `delete this` - and crashes stopped.
I find this hacky, so the caller now deletes it.

I also modified the close others, and close all functions to use
existing code.

What happened: the client was deleted (later). Then the tab index was
changed, to the deleted client, which caused a dangling pointer. With
`delete this`, the deletion is done in synchroneous mode, and it no
longer crashes here. This used to work in previous version. I think this
broke in Qt 6.10.x, as verions compiled with 6.9 do work.

closes #30